feat: finished multiDateInput implementation
