+++ apps/admin-preset-default/src/__tests__/testComponents.spec.ts
@@ -37,7 +37,7 @@ import {
   DefaultTimeInput,
   DefaultToggleInput,
   DefaultTriStateInput,
-  DefaultMultiDatesInput,
+  DefaultMultiDateInput,
 } from '../components';
 
 executePdkComponentTests({
@@ -78,5 +78,5 @@ executePdkComponentTests({
   [AdminComponent.TimeInput]: DefaultTimeInput,
   [AdminComponent.ToggleInput]: DefaultToggleInput,
   [AdminComponent.TriStateInput]: DefaultTriStateInput,
-  [AdminComponent.MultiDatesInput]: DefaultMultiDatesInput,
+  [AdminComponent.MultiDateInput]: DefaultMultiDateInput,
 });
