+++ apps/admin-preset-default/src/components/index.ts
@@ -34,7 +34,7 @@ export {default as DefaultLoader} from './DefaultLoader.vue';
 
 export {default as DefaultModal} from './DefaultModal.vue';
 
-export {default as DefaultMultiDatesInput} from './DefaultMultiDatesInput.vue';
+export {default as DefaultMultiDateInput} from './DefaultMultiDateInput.vue';
 
 export {default as DefaultMultiSelectInput} from './DefaultMultiSelectInput.vue';
