+++ package.json
@@ -94,8 +94,5 @@
         "path": "apps/checkout-js/dist/**/*.js"
       }
     ]
-  },
-  "dependencies": {
-    "@vuepic/vue-datepicker": "^11.0.2"
   }
 }
